added warning dialogue box
warning dialogue box before deleting a topic

+++ lib/widgets/topics.dart
@@ -38,8 +38,34 @@ class TopicsState extends State<Topics> {
   }
 
   Future<void> deleteItem(int id) async {
-    await dbHelper.deleteTopic(id);
-    _loadTopics();
+    bool confirmDelete = await showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text('Confirm Deletion'),
+          content: Text('Are you sure you want to delete this item?'),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(true);
+              },
+              child: Text('Delete'),
+            ),
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(false);
+              },
+              child: Text('Cancel'),
+            ),
+          ],
+        );
+      },
+    );
+
+    if (confirmDelete == true) {
+      await dbHelper.deleteTopic(id);
+      _loadTopics();
+    }
   }
 
   @override
